+++ wowaddon.py
@@ -10,16 +10,16 @@ import subprocess
 import sys
 import zipfile
 
-VERSION = '2022.2.6'  # year.month.build_num
+VERSION = '2022.7.0'  # year.month.build_num
 
-UI_VERSION_CLASSIC = '11401'  # patch 1.14.1
-BOM_NAME = 'Restocker'  # Directory and zip name
-BOM_NAME_CLASSIC = BOM_NAME + 'Classic'  # Directory and zip name
-BOM_TITLE_CLASSIC = BOM_NAME + ' Classic'  # Title field in TOC
+ADDON_NAME = 'Restocker'  # Directory and zip name
+ADDON_NAME_CLASSIC = ADDON_NAME  # Directory and zip name
+ADDON_TITLE_CLASSIC = ADDON_NAME  # Title field in TOC
+LEGACY_OVERRIDE_ADDON = f"{ADDON_NAME_CLASSIC}TBC" # subdirectory with dummy addon
 
-UI_VERSION_BCC = '20502'  # patch 2.5.2
-BOM_NAME_BCC = BOM_NAME + 'TBC'  # Directory and zip name
-BOM_TITLE_BCC = BOM_NAME + ' TBC'  # Title field in TOC
+UI_VERSION_CLASSIC = '11403'  # patch 1.14.3
+UI_VERSION_CLASSIC_TBC = '20504'  # patch 2.5.4 Phase 4 and 5 TBC
+UI_VERSION_CLASSIC_WOTLK = '30400'  # patch 3.4.0 WotLK
 
 COPY_DIRS = ['Frames', 'Classes', 'Src', 'Ace3']
 COPY_FILES = ['embeds.xml']
@@ -31,23 +31,24 @@ class BuildTool:
         self.version = VERSION
         self.copy_dirs = COPY_DIRS[:]
         self.copy_files = COPY_FILES[:]
-        self.create_toc(dst=BuildTool.toc_name_classic(),
+        self.create_toc(dst=f'{ADDON_NAME_CLASSIC}.toc',
                         ui_version=UI_VERSION_CLASSIC,
-                        title=BOM_TITLE_CLASSIC)
-        self.create_toc(dst=BuildTool.toc_name_bcc(),
-                        ui_version=UI_VERSION_BCC,
-                        title=BOM_TITLE_BCC)
-
-    @staticmethod
-    def toc_name_bcc():
-        return f'{BOM_NAME_BCC}.toc'
-
-    @staticmethod
-    def toc_name_classic():
-        return f'{BOM_NAME_CLASSIC}.toc'
+                        title=ADDON_TITLE_CLASSIC)
+        self.create_toc(dst=f'{ADDON_NAME_CLASSIC}-Classic.toc',
+                        ui_version=UI_VERSION_CLASSIC,
+                        title=ADDON_TITLE_CLASSIC)
+        self.create_toc(dst=f'{ADDON_NAME_CLASSIC}-BCC.toc',
+                        ui_version=UI_VERSION_CLASSIC_TBC,
+                        title=ADDON_TITLE_CLASSIC)
+        self.create_toc(dst=f'{ADDON_NAME_CLASSIC}-WOTLKC.toc',
+                        ui_version=UI_VERSION_CLASSIC_WOTLK,
+                        title=ADDON_TITLE_CLASSIC)
 
     def do_install(self, toc_name: str):
         self.copy_files.append(f'{toc_name}.toc')
+        self.copy_files.append(f'{toc_name}-Classic.toc')
+        self.copy_files.append(f'{toc_name}-BCC.toc')
+        self.copy_files.append(f'{toc_name}-WOTLKC.toc')
         dst_path = f'{self.args.dst}/{toc_name}'
 
         if os.path.isdir(dst_path):
@@ -68,26 +69,40 @@ class BuildTool:
 
     @staticmethod
     def do_zip_add_dir(zip: zipfile.ZipFile, dir: str, toc_name: str):
+        """ Add a directory to the zipfile, inside TOC_NAME/... subdir """
         for file in os.listdir(dir):
             file = dir + "/" + file
             print(f'ZIP: Directory {file}/')
             if os.path.isdir(file):
-                BuildTool.do_zip_add_dir(zip,
-                                         dir=file,
-                                         toc_name=toc_name)
+                BuildTool.do_zip_add_dir(zip, dir=file, toc_name=toc_name)
             else:
                 zip.write(file, f'{toc_name}/{file}')
 
+    @staticmethod
+    def do_zip_add_root_dir(zip: zipfile.ZipFile, dir: str, toc_name: str):
+        """ Add a directory to the root of the zip file """
+        for file in os.listdir(dir):
+            file = dir + "/" + file
+            print(f'ZIP: Directory {file}/')
+            if os.path.isdir(file):
+                BuildTool.do_zip_add_root_dir(zip, dir=file, toc_name=toc_name)
+            else:
+                zip.write(file, file)
+
     def do_zip(self, toc_name: str):
         self.copy_files.append(f'{toc_name}.toc')
-        zip_name = f'{self.args.dst}/{toc_name}-{self.version}.zip'
+        self.copy_files.append(f'{toc_name}-Classic.toc')
+        self.copy_files.append(f'{toc_name}-BCC.toc')
+        self.copy_files.append(f'{toc_name}-WOTLKC.toc')
+        zip_name = f'{self.args.dst}/{toc_name}-{VERSION}.zip'
 
         with zipfile.ZipFile(zip_name, "w", zipfile.ZIP_DEFLATED,
                              allowZip64=True) as zip_file:
+            # Add deprecation addon to zip
+            BuildTool.do_zip_add_root_dir(zip_file, dir=LEGACY_OVERRIDE_ADDON, toc_name=toc_name)
+
             for input_dir in self.copy_dirs:
-                BuildTool.do_zip_add_dir(zip_file,
-                                         dir=input_dir,
-                                         toc_name=toc_name)
+                BuildTool.do_zip_add_dir(zip_file, dir=input_dir, toc_name=toc_name)
 
             for input_f in self.copy_files:
                 print(f'ZIP: File {input_f}')
@@ -98,15 +113,16 @@ class BuildTool:
         # Call: git rev-parse HEAD
         p = subprocess.check_output(
             ["git", "rev-parse", "HEAD"])
-        hash = str(p).rstrip("\\n'").lstrip("b'")
-        return hash[:8]
+        hash1 = str(p).rstrip("\\n'").lstrip("b'")
+        return hash1[:8]
 
-    def create_toc(self, dst: str, ui_version: str, title: str):
-        hash = BuildTool.git_hash()
+    @staticmethod
+    def create_toc(dst: str, ui_version: str, title: str):
+        hash1 = BuildTool.git_hash()
 
         template = open('toc_template.toc', "rt").read()
         template = template.replace('${UI_VERSION}', ui_version)
-        template = template.replace('${VERSION}', f'{VERSION}-{hash}')
+        template = template.replace('${VERSION}', f'{VERSION}-{hash1}')
         template = template.replace('${ADDON_TITLE}', title)
 
         with open(dst, "wt") as out_f:
@@ -115,7 +131,7 @@ class BuildTool:
 
 def main():
     parser = argparse.ArgumentParser(
-        description="DropTrash Release and Install tool")
+        description="Restocker Release and Install tool")
     parser.add_argument(
         '--dst', type=str, required=True, action='store',
         help='The destination directory where the game Addons will be copied, '
@@ -123,8 +139,8 @@ def main():
              'name.')
 
     parser.add_argument(
-        '--version', choices=['classic', 'tbc'],
-        help='The version to copy or zip, classic or TBC')
+        '--version', choices=['classic', 'tbc', 'wotlk'],
+        help='The version to copy or zip: classic, TBC or WotLK')
 
     parser.add_argument(
         'command', choices=['help', 'zip', 'install'],
@@ -136,11 +152,11 @@ def main():
 
     if args.command == 'install':
         bt = BuildTool(args)
-        bt.do_install(toc_name=BOM_NAME_CLASSIC if args.version == 'classic' else BOM_NAME_BCC)
+        bt.do_install(toc_name=ADDON_NAME_CLASSIC)
 
     elif args.command == 'zip':
         bt = BuildTool(args)
-        bt.do_zip(toc_name=BOM_NAME_CLASSIC if args.version == 'classic' else BOM_NAME_BCC)
+        bt.do_zip(toc_name=ADDON_NAME_CLASSIC)
     else:
         parser.print_help()